requirements.txt
future
emoji
requests
sqlalchemy
python-telegram-bot>=11.1.0
psycopg2
feedparser
pyowm
pythonping
pynewtonmath
#memes from SkittBot
spongemock
zalgo-text
geopy
nltk
aiohttp>=2.2.5
Pillow>=4.2.0
#Cool stuff by Peak
PyLyrics
#AEX
hurry.filesize
#translator
googletrans

tldextract
